Partial work for issue 654. Not finished but partially working.

+++ core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/OClusterLocalHole.java
@@ -110,7 +110,7 @@ public class OClusterLocalHole extends OSingleFileSegment {
 	 * @return the recycled position if found, otherwise -1 that usually means to request more space.
 	 * @throws IOException
 	 */
-	public long getEntryPosition(final long iPosition) throws IOException {
+	public long getEntryPosition(final int iPosition) throws IOException {
 		return file.readLong(iPosition * RECORD_SIZE);
 	}
